requirements.txt
pydantic>=2.0
sentry-sdk
pandas>=2.0
scikit-learn>=1.3
Jinja2>=3.1
PyYAML>=6.0
# Optional: tiktoken (for precise token counting in adapter builds)
# tiktoken>=0.5
# Optional: APScheduler (for report scheduler - recommended)
# apscheduler>=3.10
